refactor: extract seeding logic into seed package (#10)
